Rename backstack package to configuration
