+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/configuration/action/multi/MultiConfigurationAction.kt
@@ -1,10 +1,10 @@
-package com.badoo.ribs.core.routing.backstack.action.multi
+package com.badoo.ribs.core.routing.configuration.action.multi
 
 import android.os.Parcelable
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Resolved
-import com.badoo.ribs.core.routing.backstack.ConfigurationKey
-import com.badoo.ribs.core.routing.backstack.action.ActionExecutionParams
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.Resolved
+import com.badoo.ribs.core.routing.configuration.ConfigurationKey
+import com.badoo.ribs.core.routing.configuration.action.ActionExecutionParams
 
 /**
  * Represents an action that affects multiple configurations when executed.
