+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/configuration/feature/ConfigurationFeature.kt
@@ -1,4 +1,4 @@
-package com.badoo.ribs.core.routing.backstack.feature
+package com.badoo.ribs.core.routing.configuration.feature
 
 import android.os.Parcelable
 import com.badoo.mvicore.element.Actor
@@ -8,28 +8,28 @@ import com.badoo.mvicore.element.TimeCapsule
 import com.badoo.mvicore.feature.ActorReducerFeature
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.routing.action.RoutingAction
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand.SaveInstanceState
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand.Sleep
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand.WakeUp
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Activate
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Add
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Deactivate
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Remove
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.ActivationState.ACTIVE
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.ActivationState.INACTIVE
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.ActivationState.SLEEPING
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Resolved
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Unresolved
-import com.badoo.ribs.core.routing.backstack.ConfigurationKey
-import com.badoo.ribs.core.routing.backstack.action.ActionExecutionParams
-import com.badoo.ribs.core.routing.backstack.action.single.AddAction
-import com.badoo.ribs.core.routing.backstack.action.multi.MultiConfigurationAction
-import com.badoo.ribs.core.routing.backstack.action.single.SingleConfigurationAction
-import com.badoo.ribs.core.routing.backstack.feature.ConfigurationFeature.Effect
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand.SaveInstanceState
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand.Sleep
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand.WakeUp
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Activate
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Add
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Deactivate
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Remove
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.ActivationState.ACTIVE
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.ActivationState.INACTIVE
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.ActivationState.SLEEPING
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.Resolved
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.Unresolved
+import com.badoo.ribs.core.routing.configuration.ConfigurationKey
+import com.badoo.ribs.core.routing.configuration.action.ActionExecutionParams
+import com.badoo.ribs.core.routing.configuration.action.single.AddAction
+import com.badoo.ribs.core.routing.configuration.action.multi.MultiConfigurationAction
+import com.badoo.ribs.core.routing.configuration.action.single.SingleConfigurationAction
+import com.badoo.ribs.core.routing.configuration.feature.ConfigurationFeature.Effect
 import io.reactivex.Observable
 import io.reactivex.Observable.empty
 import io.reactivex.Observable.fromCallable
