+++ android/libraries/rib-base/src/test/java/com/badoo/ribs/core/routing/configuration/ConfigurationFeatureTest.kt
@@ -1,4 +1,4 @@
-package com.badoo.ribs.core.routing.backstack
+package com.badoo.ribs.core.routing.configuration
 
 import android.os.Bundle
 import android.os.Parcelable
@@ -7,27 +7,27 @@ import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.Node.Descriptor
 import com.badoo.ribs.core.Node.ViewAttachMode
 import com.badoo.ribs.core.routing.action.RoutingAction
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand.Sleep
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.MultiConfigurationCommand.WakeUp
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Activate
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Add
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Deactivate
-import com.badoo.ribs.core.routing.backstack.ConfigurationCommand.SingleConfigurationCommand.Remove
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.ActivationState.INACTIVE
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.ActivationState.SLEEPING
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Resolved
-import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Unresolved
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.ContentExternal1
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.ContentExternal2
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.ContentViewParented1
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.ContentViewParented2
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.ContentViewParented3
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.Permanent1
-import com.badoo.ribs.core.routing.backstack.ConfigurationFeatureTest.Configuration.Permanent2
-import com.badoo.ribs.core.routing.backstack.ConfigurationKey.Content
-import com.badoo.ribs.core.routing.backstack.ConfigurationKey.Permanent
-import com.badoo.ribs.core.routing.backstack.feature.ConfigurationFeature
-import com.badoo.ribs.core.routing.backstack.feature.SavedState
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand.Sleep
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.MultiConfigurationCommand.WakeUp
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Activate
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Add
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Deactivate
+import com.badoo.ribs.core.routing.configuration.ConfigurationCommand.SingleConfigurationCommand.Remove
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.ActivationState.INACTIVE
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.ActivationState.SLEEPING
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.Resolved
+import com.badoo.ribs.core.routing.configuration.ConfigurationContext.Unresolved
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.ContentExternal1
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.ContentExternal2
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.ContentViewParented1
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.ContentViewParented2
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.ContentViewParented3
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.Permanent1
+import com.badoo.ribs.core.routing.configuration.ConfigurationFeatureTest.Configuration.Permanent2
+import com.badoo.ribs.core.routing.configuration.ConfigurationKey.Content
+import com.badoo.ribs.core.routing.configuration.ConfigurationKey.Permanent
+import com.badoo.ribs.core.routing.configuration.feature.ConfigurationFeature
+import com.badoo.ribs.core.routing.configuration.feature.SavedState
 import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.clearInvocations
 import com.nhaarman.mockitokotlin2.doAnswer
